fix 짜집기중

+++ Cronos_URP/Assets/Script/TestEnemyAI_Script/DamageSystem/MeleeWeapon.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 public class MeleeWeapon : MonoBehaviour
 {
-    public int damage = 1;
+    public float damage = 1f;
 
     [System.Serializable]
     public class AttackPoint
